requirements.txt
replicate
google-generativeai
pydub
elevenlabs==0.3.0b0